requirements.txt
aiohttp>=3.7.4
pydantic>=1.8.2
python-dotenv>=0.19.0
